-Cart Bug Fixed (#166)

+++ src/main/kotlin/com/example/data/local/table/cart/CartTable.kt
@@ -5,7 +5,7 @@ import org.jetbrains.exposed.sql.Column
 import org.jetbrains.exposed.sql.Table
 
 object CartTable: Table("Cart") {
-    val cartId: Column<Long> = long("cartId").autoIncrement()
+    val cartId: Column<Int> = integer("cartId").autoIncrement()
     val productId: Column<Long> = long("productId")
     val quantity: Column<Int> = integer("quantity")
     val userId: Column<Long> = long("userId").references(UserTable.id)
