+++ src/main/kotlin/com/example/data/repository/cart/CartDao.kt
@@ -11,17 +11,17 @@ interface CartDao {
     suspend fun getAllCart(): List<CartItem>?
     suspend fun getCartByUserId(id: Long): List<CartItem>?
     suspend fun getCartItemByUserId(id: Long): CartItem?
-    suspend fun getCartItemByCartId(id: Long): CartItem?
-    suspend fun deleteCartItemByCartId(id:Long): Int?
+    suspend fun getCartItemByCartId(id: Int): CartItem?
+    suspend fun deleteCartItemByCartId(id:Int): Int?
     suspend fun deleteCartByUserId(id:Long): Int?
     suspend fun update(
-        cartId: Long,
+        cartId: Int,
         productId: Long,
         quantity: Int,
         userId: Long
     )
     suspend fun updateCartItem(
-        cartId: Long,
+        cartId: Int,
         productId: Long,
         quantity: Int,
         userId: Long
